use response from client data store to update vuex store
the untrust device function now returns a response which is passed to the vuex store to update local state.
previously, the store action handled the specifics of how store and client data should look. the new setup is less brittle, as if the data needed to ascertain trust changes, it will now be handled in one place