enable url navigation
navigation via the url was not working, as the user defaulted to unauthenticated. this meant that when the page was refreshed, the logged in state was false which meant that the router redirected the navigation to the home page. fixed by moving the re-route command into the sign out method, instead of the auth state listener, and by defaulting the user to logged in until the listener is triggered.