moved redirect back to auth state listener
the previous commit removed the redirect functionality into the function triggered by the signing out, rather than the auth state listener. This allowed all unauthenticated users to navigate to any page via the url. moving the redirect function back to the listener and leaving the user defaulted to logged in allows the user to navigate using the url when logged in, but forces redirect to home while not logged in.