enabled firestore
This was a particularly difficult feature to implement. Google reccommend asking before enabling offline persistence, so i had to make some way of marking the device as trusted. I then had to find out if the device is trusted before enabling firestore. This proved difficult as for some reason, when i tried to import the Vuex store into the module that initialized firestore, the local store failed to initialize. i fixed this by reading from the data in the cache, rather than in memory.

The next issue is that I was testing if the device was trusted before making any request to the database. this was actually not allowed by firebase. it turns out that persistence must be enabled at app start and cannot be disabled again. I now check if the device is trusted at app start, and then enable persistence if so.