generified add income view
moved the logic for creating finance data from the AddFinanceFormWrapper to a specific parent component. this allowed the form wrapper to become a generic component that just emits the data to the parent and the specific parent decides on what should be done with the data collected from the form.