moved vuex store modules from uiComponents
Moved the local auth module from the uiComponents directory out to the store module. The reason for this decision was that it made more sense, since the application store holds the data to be shared across components, that the modules be accessed from the data directory instead.

The is also a planned refactor to export the type names used to dispatch/commit to the store. having all of the store modules in the same place allows easy exporting as a single object.

This was a reasonably short refactor. for the most part it just involved moving the files and then using find/replace across the project directory to update the import paths. this also streamlined the auth ui module, as it now no longer had to export anything. I had to slightly update the barrel imports/exports to reflect the new directory structure.