restructured type exports
the types definition file for working with vuex store modules used to have to be imported from each individual module. now they get exported up to the top level of the app directory.

this means that this file can be more easily imported using a much shorter, aliased, path. This also means that modules that must access multiple store modules will have access to all type definitions from a single import.

In order to do this refactor the index.js file in each store module had to be modified to export the type definitions as well as the store definition. this export is continued to up to the app level index.js, where it is split out into two seperate imports. this allows the store to import the store definition from here, while other modules can access the type definitions from the same location.

I also changed all of the paths throughout the application to import the type definitions from the new locations using find and replace.