moved namespace function to store directory
the namspace function is a utility to prefix the types of store modules with their respective module name. since this file is only required by the type definition files for each module, it made sense to remove it from the app level util directory and place it into the app/data/store directory. this allowed me to remove the app/util directory, and slightly flatten the application structure.