created http utils
created the http utils module. the main purpose for this module at the minute is to generate the authentication header for http requests. This module will also be used in the future to house other utilities specific to http requests.

this module also provides an abstraction for fetching a users id. since the user id does not exist immediatly following a page reload, this abstraction will attempt to fetch the id 20 times, waiting 50ms between each fetch. this will give firebase time to log in the user in the event of page reloads.

may refactor this functionality so that it is performed in the auth service instead.