Added new project dependencies
Added two new dependencies to the project (vue-currency-filter and date-fns).

Vue-Currency-Filter is a vue filter that allows numerical data to be formatted as currency. This will be useful as there will be a lot of financial data to be displayed.

Date-Fns is a lightweight date formatting library that seeks to overcome the inherent problems with native JS Date objects. This is a fairly new library on the scene which seeks to improve on the current offerings (such as momentjs) by adhering to a functional paradigm, as well as a function-per-file build strategy. The build strategy helps to reduce bloat as each function is completley standalone and must be imported from seperate files, thus reducing vendor bundle size. It also takles the problem of mutability that was such an issue with momentjs through the use of the functional paradigm.