Created Long Date filter
Created a Vue filter designed to convert the dates stored in the db into a more human readable format. This is done using the date-fns library imported previously. This filter will convert dates in the format of "YYYY-MM-DD" to "dddd Do MMMM, YYYY". An example ofthis is converting "2018-02-19" to "Monday 19th February, 2018"