Created generic display finance component
Created a generic data table that will be used to display finance details. This table is will be used to display incomes, expenditures or both together.

The table comes on top of a card which contains a title and search bar. All columns of the table are sortable.  The table has an optional extra column to display the type of the transaction when it is not displaying a single type.