Created generic card for overview of financial details.
Created a card component which can be used to display an overview of financial details. This card is designed to be displayed in conjunction with the table designed earlier. The card will display details pertaining to the current transactions, such as totals, averages and dates.

This card is designed to be used with all trasnaction types, so it was quite difficult to build. It is a generic component which displays different data depending on if a since transaction type is shown, or if both income and expenditure is shown. For example, if it is currently just showing incomes, there will be a single total field, but if both incomes and expenditires are being shown there will be a total income and total expenditure field.

This card also contains actions for interacting with the data, such as fetching more records and adding new records.

There is also a slot for advanced searching/filtering options, but these have yet to be implemented.