Created finance filter component
Created a finance filter component that will be used to select dates for filtering displayed financial transactions. This component comtains two date pickers, one for date and one for month as well as a radio to select which picker is currently visible. When a date is selected, the value is emitted to the parent to be used for filtering