Created summary stats componenents
Created summary stats components that can be used to display the summary stats about the currently viewed transactions. The current transactions are passed into these components as props, so thes components have nothing to do with the filtering of the data.

There are two components, one for displaying stats about a single type of transaction and one for displaying mixed records about both types of transactions