Restructured ManageFinanceCard
Restructured the manage finance card to make use of the new filter component and stats component. The markup and logic for these two child components was originally contained within this component, which led to the component being bloated. Now this component handles how these children should be displayed as a single entity