Added new filter options
Handle filtering based on the data that bubbles up from the filter component. This component acts as the point where data is injected into the finace component tree and is used to handle all filtering for the data. So in this case, the manage finance card will emit the filter data, this component will filter the list of data and then pass the result back to both the table and the manage finance card.