Created add item ui
Created the UI to add an item. In this context an item is something that can be purchased, either sporadically or in a recurring manner.

This commit contains a basic UI for adding an item as well as the nessecary updates to the routing files to handle routing for this component.