Updated login form UI
Updated the login form UI to display the some text to switch between login and signup mode. This is more in line with common conventions than the checkbox that was previously used.

This was a simple refactor. The checkbox was hidden and two new paragraphs were conditionaly shown based on the value that was previously connected to the checkbox. The values are updated by an a tag inside the paragrah of text.