Created cookie warning.
Created a snackbar that will display an informational message about the use of cookies in the application. This feature will also persist the factthat a user hass closed the cookie warning, if on a trusted device. If this is a trusted device, then the warning will not be displayed again, until the user signs out of the application. If this is a trusted device, then the warning will be displayed every time the user logs in.

This commit also took in all of the technical details associated with this feature, such as writing the cookie warning vaiable out locally and reading it into the Vuex store at application start.