Added trusted device field to login/signup
Added a trusted device field to the login/signup form. This will allow the user to mark this device as trusted, so that data can be persisted offline. This commit took in the details of adding the option to the UI, as well as the logic to persist the result to the client machine.