Improved the user id fetch
Improved the user id fetch functionality to be a bit more robust in when a user is not logged in.

Now the fetchID function will call a getUser function, which will attempt to fetch the currently logged in user 10 times, with a 50ms wait between each attempt. This should allow sufficent time for the firebase auth state change listener to be triggered if auth is in progress.

The main purpose of this refactor was to solve a bug in the transaction views. When a view that required transactional data was refreshed, an attempt was made to fetch the currently logged in user from one of the components. At this point the currently logged in user did not exist as the firebase auth event triggered by the page reload had not yet completed.

This refactor also took in the changes in the financeService module, which were made to accomodate the new feature