Refactored view finance
Refactored the layout of the view finanace page. This driver for this change was mainly the new navigation. This allowed me to remove presentation logic from this view, and listen for events from the navbar instead. This means that the accordian that used to show the stats is gone, and the contents of it are shown provisionaly based on the contents of the navbar.