Updated manifest
Updated the manifest.json file to change the color, as well as the start url. The background will now be white, and the theme will be a sarker shade of blue than the toolbar. The start url will now point to the home route '/' rather than the non-existent '/index.html' route.