Updated reference view
Updated the reference view so that there is now a seperate component used to display a single card in the data iterator. Also created a JSON file that stores the details about the references, so that the view can be built. There was also a need to add some logos to the application, to show the tools used. A link to the references was also added to the footer, to be available throughout the application.