Fixed contextual side nav display issues on mobile
Fixed the issue of the contextual side navi not displaying properly on smaller displays. This was caused by the mini variant of the side navigation being superimposed over the page content, as well as the screen being too small for the  both the content and the navigation to be displayed.

To fix this I added a resize listener that converts the sidenav to the temporary variant on anything smaller than a laptop. I also added a button that can be used to trigger the side nav that will only be visible on smaller screens. The temporary nav drawer is not visible at all when it is closed, and appears overlayed on the page content with a darkened scrim background, instead of beside it.